Improved data types and memory usage

+++ video/agon_keyboard.h
@@ -1,7 +1,6 @@
 #ifndef AGON_KEYBOARD_H
 #define AGON_KEYBOARD_H
 
-#include <Arduino.h>
 #include <fabgl.h>
 
 #include "agon.h"
@@ -9,10 +8,10 @@
 
 fabgl::PS2Controller		_PS2Controller;		// The keyboard class
 
-byte			_keycode = 0;					// Last pressed key code
-byte			_modifiers = 0;					// Last pressed key modifiers
-int				kbRepeatDelay = 500;			// Keyboard repeat delay ms (250, 500, 750 or 1000)		
-int				kbRepeatRate = 100;				// Keyboard repeat rate ms (between 33 and 500)
+uint8_t			_keycode = 0;					// Last pressed key code
+uint8_t			_modifiers = 0;					// Last pressed key modifiers
+uint16_t		kbRepeatDelay = 500;			// Keyboard repeat delay ms (250, 500, 750 or 1000)		
+uint16_t		kbRepeatRate = 100;				// Keyboard repeat rate ms (between 33 and 500)
 
 // Get keyboard instance
 inline fabgl::Keyboard* getKeyboard() {
@@ -31,7 +30,7 @@ void setupKeyboard() {
 
 // Set keyboard layout
 //
-void setKeyboardLayout(int region) {
+void setKeyboardLayout(uint8_t region) {
 	auto kb = getKeyboard();
 	switch(region) {
 		case 1:	kb->setLayout(&fabgl::USLayout); break;
@@ -51,7 +50,7 @@ void setKeyboardLayout(int region) {
 // Get keyboard key presses
 // returns true only if there's a new keypress update
 //
-bool getKeyboardKey(byte *keycode, byte *modifiers, byte *vk, byte *down) {
+bool getKeyboardKey(uint8_t *keycode, uint8_t *modifiers, uint8_t *vk, uint8_t *down) {
 	auto kb = getKeyboard();
 	fabgl::VirtualKeyItem item;
 
@@ -117,7 +116,7 @@ bool getKeyboardKey(byte *keycode, byte *modifiers, byte *vk, byte *down) {
 
 // Simpler keyboard read for CP/M Terminal Mode
 //
-bool getKeyboardKey(byte *ascii) {
+bool getKeyboardKey(uint8_t *ascii) {
 	auto kb = getKeyboard();
 	fabgl::VirtualKeyItem item;
 
@@ -135,7 +134,7 @@ bool getKeyboardKey(byte *ascii) {
 
 // Wait for shift key to be released, then pressed (used for paged mode)
 // 
-bool wait_shiftkey(byte *ascii, byte* vk, byte* down) {
+bool wait_shiftkey(uint8_t *ascii, uint8_t* vk, uint8_t* down) {
 	auto kb = getKeyboard();
 	fabgl::VirtualKeyItem item;
 
@@ -160,7 +159,7 @@ bool wait_shiftkey(byte *ascii, byte* vk, byte* down) {
 	return true;
 }
 
-void getKeyboardState(int *repeatDelay, int *repeatRate, byte *ledState) {
+void getKeyboardState(uint16_t *repeatDelay, uint16_t *repeatRate, uint8_t *ledState) {
 	bool numLock;
 	bool capsLock;
 	bool scrollLock;
@@ -170,7 +169,7 @@ void getKeyboardState(int *repeatDelay, int *repeatRate, byte *ledState) {
     *repeatRate = kbRepeatRate;
 }
 
-void setKeyboardState(int delay, int rate, byte ledState) {
+void setKeyboardState(uint16_t delay, uint16_t rate, uint8_t ledState) {
 	auto kb = getKeyboard();
 	if (delay >= 250 && delay <= 1000) kbRepeatDelay = (delay / 250) * 250;	// In steps of 250ms
 	if (rate >=  33 && rate <=  500) kbRepeatRate  = rate;
